requirements.txt
langchain
python-dotenv
streamlit
openai
Tabulate 
plotly
altair